Update check-vpn-speeds.sh
A lot of changes going on here. I've modified the trailing slash issues. Also, one of the biggest things plaguing this script was that issue when a .conf file left the script hanging when a valid VPN connection couldn't be made. I've implemented a counter with the "TIMEOUT" variable up top so that after issuing the command to start the VPN, if it doesn't start within that timeout period, it moves on to the next connection. A small aesthetic change to also tell you which VPN file it is on out of how many are in the directory total. I've cleaned up a little bit of code here and there, as well.