Update check-vpn-speeds.sh
I finally got the final piece of the puzzle worked out, and got a piece implemented that I really wanted. Now, when cycling through VPN files, it will ping the remote address in the .conf file. If the average of the 4 ping requests is higher than the PINGTHRESHOLD variable at the top, that VPN conf file will be skipped and the next one loaded. You can modify the threshold you're willing to work with at the top, but I find that 100 provides a good limit on which servers to skip and which to keep using.